requirements.txt
amqp==2.4.2
anyjson==0.3.3
billiard==3.5.0.5
celery==4.2.2
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cmake==3.13.3
cycler==0.10.0
dj-static==0.0.6
Django==2.1.2
django-background-tasks==1.1.13
django-compat==1.0.15
django-environ==0.4.1
django-extra-fields==1.0.0
django-rq==1.3.0
djangorestframework==3.9.2
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
gunicorn==19.9.0
idna==2.8
jsonfield==2.0.2
kiwisolver==1.0.1
kombu==4.2.0
netifaces==0.10.9
numpy==1.16.2
Pillow==5.4.1
ply==3.11
psycopg2==2.7.7
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.4
redis==2.10.6
requests==2.21.0
rq==0.12.0
six==1.11.0
static3==0.7.0
urllib3==1.24.1
vine==1.3.0
whitenoise==4.1.2
widgetsnbextension==3.4.2